Removed _private_tools module. Exceptions where moved to different files in the modules which they are used
