+++ openpharmacophore/point/input_arguments.py
@@ -1,4 +1,4 @@
-from openpharmacophore._private_tools.exceptions import IncorrectShapeError, NotArrayLikeError, \
+from openpharmacophore.point.exceptions import IncorrectShapeError, NotArrayLikeError, \
     NotAQuantityError, WrongDimensionalityError
 import numpy as np
 import pyunitwizard as puw
